refactor: split MainWindow into partial classes
